Remove duplicated navbar
